refactor: echo file 분리
